REFACT :: DayOfWeek을 영어로도 입력 가능하도록 수정
