+++ maeumgagym-presentation/src/main/kotlin/com/info/maeumgagym/controller/routine/dto/UpdateRoutineWebRequest.kt
@@ -1,6 +1,6 @@
 package com.info.maeumgagym.controller.routine.dto
 
-import com.info.maeumgagym.common.convertor.DayOfWeekKoreanConvertor
+import com.info.maeumgagym.common.convertor.DayOfWeekConvertor
 import com.info.maeumgagym.routine.dto.request.ExerciseInfoRequest
 import com.info.maeumgagym.routine.dto.request.UpdateRoutineRequest
 import javax.validation.constraints.NotBlank
@@ -30,7 +30,7 @@ data class UpdateRoutineWebRequest(
         isShared!!,
         exerciseInfoRequestList!!,
         dayOfWeeks!!.map {
-            DayOfWeekKoreanConvertor.koreanToDayOfWeek(it)
+            DayOfWeekConvertor.stringToDayOfWeek(it)
         }.toMutableSet()
     )
 }
